+++ checkout-ui-custom/checkout6-custom.js
@@ -246,7 +246,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \************************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _ContactCard__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./ContactCard */ \"./src/partials/Deliver/ContactCard.js\");\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./utils */ \"./src/partials/Deliver/utils.js\");\n\n\n\nvar AddAddressAutoCompleteForm = function AddAddressAutoCompleteForm(address) {\n  var fields = [{\n    name: 'number',\n    required: true,\n    label: 'Street number',\n    value: ''\n  }, {\n    name: 'addressType',\n    label: 'Address type',\n    required: true,\n    type: 'radio',\n    options: [{\n      value: 'residential',\n      label: 'Residential',\n      checked: true\n    }, {\n      value: 'business',\n      label: 'Business'\n    }]\n  }, {\n    name: 'businessName',\n    label: 'Business name',\n    required: false,\n    value: '',\n    maxLength: 100\n  }, {\n    name: 'companyBuilding',\n    label: 'Building/Complex and number',\n    required: false,\n    value: '',\n    maxLength: 100\n  }, {\n    name: 'receiverName',\n    label: 'Recipient’s name',\n    required: true,\n    value: (0,_utils__WEBPACK_IMPORTED_MODULE_2__.getBestRecipient)({\n      type: 'delivery'\n    })\n  }, {\n    name: 'complement',\n    // TODO: ask john what is this?\n    required: false,\n    type: 'hidden',\n    helperText: 'We send shipping updates to this number.',\n    value: ''\n  }, {\n    name: 'receiverPhone',\n    label: 'Recipient’s mobile number',\n    required: true,\n    type: 'tel',\n    helperText: 'We send shipping updates to this number.',\n    minlength: 9,\n    error: 'Please enter a valid phone number',\n    containerClasses: 'custom-field-complement' // for sa flag\n  }];\n\n  // TODO: VALIDATE DATA RETURNED FROM  GOOGLE - IS IT SUFFICIENT\n  // TODO: IF YES - MOVE TO THIS FORM\n  // TODO: Rename  ID FORM\n  // TODO: Add on SUBMIT\n  // TODO: LEGIT WHERE ARE WE STORING THIS INFO FOR  ON SUBMIT\n  // TODO: ASK JOHN WHAT IS COMPLEMENT\n  // TODO: How to deal with seperate street number  field\n  return (/* html */\"\\n    <form id=\\\"bash--address-form\\\" method=\\\"post\\\">\\n      \".concat((0,_ContactCard__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(address), \"\\n      \").concat(fields.map(function (field) {\n      return (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_1__[\"default\"])(field);\n    }).join(''), \"\\n      <button\\n        class=\\\"submit btn-go-to-payment btn btn-large btn-success\\\"\\n        id=\\\"btn-save-address\\\"\\n        type=\\\"submit\\\">\\n        Save address\\n      </button>\\n    </form>\\n  \")\n  );\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (AddAddressAutoCompleteForm);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/AddAddressAutoCompleteForm.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\n/* harmony import */ var _FormComponents__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./FormComponents */ \"./src/partials/Deliver/FormComponents.js\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./utils */ \"./src/partials/Deliver/utils.js\");\n\n\n\nvar AddAddressAutoCompleteForm = function AddAddressAutoCompleteForm(address) {\n  var fields = [{\n    name: 'number',\n    required: true,\n    label: 'Street number',\n    value: ''\n  }, {\n    name: 'addressType',\n    label: 'Address type',\n    required: true,\n    type: 'radio',\n    options: [{\n      value: 'residential',\n      label: 'Residential',\n      checked: true\n    }, {\n      value: 'business',\n      label: 'Business'\n    }]\n  }, {\n    name: 'businessName',\n    label: 'Business name',\n    required: false,\n    value: '',\n    maxLength: 100\n  }, {\n    name: 'companyBuilding',\n    label: 'Building/Complex and number',\n    required: false,\n    value: '',\n    maxLength: 100\n  }, {\n    name: 'receiverName',\n    label: 'Recipient’s name',\n    required: true,\n    value: (0,_utils__WEBPACK_IMPORTED_MODULE_2__.getBestRecipient)({\n      type: 'delivery'\n    })\n  }, {\n    name: 'complement',\n    // TODO: ask john what is this?\n    required: false,\n    type: 'hidden',\n    helperText: 'We send shipping updates to this number.',\n    value: ''\n  }, {\n    name: 'receiverPhone',\n    label: 'Recipient’s mobile number',\n    required: true,\n    type: 'tel',\n    helperText: 'We send shipping updates to this number.',\n    minlength: 9,\n    error: 'Please enter a valid phone number',\n    containerClasses: 'custom-field-complement' // for sa flag\n  }];\n\n  // TODO: VALIDATE DATA RETURNED FROM  GOOGLE - IS IT SUFFICIENT\n  // TODO: IF YES - MOVE TO THIS FORM\n  // TODO: Rename  ID FORM\n  // TODO: Add on SUBMIT\n  // TODO: LEGIT WHERE ARE WE STORING THIS INFO FOR  ON SUBMIT\n  // TODO: ASK JOHN WHAT IS COMPLEMENT\n  // TODO: How to deal with seperate street number  field\n  return (/* html */\"\\n    \".concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_1__.AddressSectionHeading)('Delivery Details'), \"\\n    \").concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_1__.ContactCard)(address), \"\\n    <form id=\\\"bash--address-form\\\" method=\\\"post\\\">\\n      \").concat(fields.map(function (field) {\n      return (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(field);\n    }).join(''), \"\\n      \").concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_1__.SubmitButton)(), \"\\n    </form>\\n  \")\n  );\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (AddAddressAutoCompleteForm);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/AddAddressAutoCompleteForm.js?");
 
 /***/ }),
 
@@ -300,16 +300,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/partials/Deliver/ContactCard.js":
-/*!*********************************************!*\
-  !*** ./src/partials/Deliver/ContactCard.js ***!
-  \*********************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/**\n * ContactCard - View Component of a Address Summary\n * @param {Object} data - Address data\n */\nvar ContactCard = function ContactCard(data) {\n  var number = data.number,\n    businessName = data.businessName,\n    street = data.street,\n    neighborhood = data.neighborhood,\n    city = data.city,\n    postalCode = data.postalCode;\n  var addressLine = [businessName, \"\".concat(number ? \"\".concat(number, \" \") : '').concat(street), neighborhood !== null && neighborhood !== void 0 ? neighborhood : city, postalCode].filter(function (item) {\n    return item !== undefined && item !== null && item !== '';\n  }).join(', ').trim();\n  return (/* html */\"\\n    <div style=\\\"border: 0.5px solid #0404041A; border-radius: 8px; padding: 16px; display: flex; align-items: center;\\\">\\n      <svg width=\\\"24\\\" height=\\\"19\\\" viewBox=\\\"0 0 24 19\\\" fill=\\\"none\\\" xmlns=\\\"http://www.w3.org/2000/svg\\\">\\n        <path d=\\\"M1.5 18.2496H22.5M14.25 18.2496V12.2496H9.75V18.2496M3.75 8.9393V18.2496M20.25 18.2496V8.9393M2.25 10.4393L11.4694 1.21899C11.539 1.14926 11.6217 1.09394 11.7128 1.05619C11.8038 1.01845 11.9014 0.999023 12 0.999023C12.0986 0.999023 12.1962 1.01845 12.2872 1.05619C12.3783 1.09394 12.461 1.14926 12.5306 1.21899L21.75 10.4393\\\" stroke=\\\"black\\\" stroke-width=\\\"1.5\\\" stroke-linecap=\\\"round\\\" stroke-linejoin=\\\"round\\\"/>\\n      </svg>\\n      <p style=\\\"margin-left: 16px; line-height: 1; font-size: 14px; margin-bottom:0;\\\">\".concat(addressLine, \"</p>\\n  </div>\\n  \")\n  );\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (ContactCard);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/ContactCard.js?");
-
-/***/ }),
-
 /***/ "./src/partials/Deliver/DeliverContainer.js":
 /*!**************************************************!*\
   !*** ./src/partials/Deliver/DeliverContainer.js ***!
@@ -346,7 +336,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \*************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"EDIT_FORM_RECEIVER_PHONE_ID\": () => (/* binding */ EDIT_FORM_RECEIVER_PHONE_ID),\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__),\n/* harmony export */   \"submitEditAddressForm\": () => (/* binding */ submitEditAddressForm)\n/* harmony export */ });\n/* harmony import */ var libphonenumber_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! libphonenumber-js */ \"./node_modules/libphonenumber-js/index.es6.exports/isValidNumber.js\");\n/* harmony import */ var _utils_phoneFields__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/phoneFields */ \"./src/utils/phoneFields.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _utils_setAddress__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/setAddress */ \"./src/utils/setAddress.js\");\n/* harmony import */ var _ContactCard__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./ContactCard */ \"./src/partials/Deliver/ContactCard.js\");\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\nfunction _regeneratorRuntime() { \"use strict\"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return exports; }; var exports = {}, Op = Object.prototype, hasOwn = Op.hasOwnProperty, defineProperty = Object.defineProperty || function (obj, key, desc) { obj[key] = desc.value; }, $Symbol = \"function\" == typeof Symbol ? Symbol : {}, iteratorSymbol = $Symbol.iterator || \"@@iterator\", asyncIteratorSymbol = $Symbol.asyncIterator || \"@@asyncIterator\", toStringTagSymbol = $Symbol.toStringTag || \"@@toStringTag\"; function define(obj, key, value) { return Object.defineProperty(obj, key, { value: value, enumerable: !0, configurable: !0, writable: !0 }), obj[key]; } try { define({}, \"\"); } catch (err) { define = function define(obj, key, value) { return obj[key] = value; }; } function wrap(innerFn, outerFn, self, tryLocsList) { var protoGenerator = outerFn && outerFn.prototype instanceof Generator ? outerFn : Generator, generator = Object.create(protoGenerator.prototype), context = new Context(tryLocsList || []); return defineProperty(generator, \"_invoke\", { value: makeInvokeMethod(innerFn, self, context) }), generator; } function tryCatch(fn, obj, arg) { try { return { type: \"normal\", arg: fn.call(obj, arg) }; } catch (err) { return { type: \"throw\", arg: err }; } } exports.wrap = wrap; var ContinueSentinel = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var IteratorPrototype = {}; define(IteratorPrototype, iteratorSymbol, function () { return this; }); var getProto = Object.getPrototypeOf, NativeIteratorPrototype = getProto && getProto(getProto(values([]))); NativeIteratorPrototype && NativeIteratorPrototype !== Op && hasOwn.call(NativeIteratorPrototype, iteratorSymbol) && (IteratorPrototype = NativeIteratorPrototype); var Gp = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(IteratorPrototype); function defineIteratorMethods(prototype) { [\"next\", \"throw\", \"return\"].forEach(function (method) { define(prototype, method, function (arg) { return this._invoke(method, arg); }); }); } function AsyncIterator(generator, PromiseImpl) { function invoke(method, arg, resolve, reject) { var record = tryCatch(generator[method], generator, arg); if (\"throw\" !== record.type) { var result = record.arg, value = result.value; return value && \"object\" == _typeof(value) && hasOwn.call(value, \"__await\") ? PromiseImpl.resolve(value.__await).then(function (value) { invoke(\"next\", value, resolve, reject); }, function (err) { invoke(\"throw\", err, resolve, reject); }) : PromiseImpl.resolve(value).then(function (unwrapped) { result.value = unwrapped, resolve(result); }, function (error) { return invoke(\"throw\", error, resolve, reject); }); } reject(record.arg); } var previousPromise; defineProperty(this, \"_invoke\", { value: function value(method, arg) { function callInvokeWithMethodAndArg() { return new PromiseImpl(function (resolve, reject) { invoke(method, arg, resolve, reject); }); } return previousPromise = previousPromise ? previousPromise.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); } }); } function makeInvokeMethod(innerFn, self, context) { var state = \"suspendedStart\"; return function (method, arg) { if (\"executing\" === state) throw new Error(\"Generator is already running\"); if (\"completed\" === state) { if (\"throw\" === method) throw arg; return { value: void 0, done: !0 }; } for (context.method = method, context.arg = arg;;) { var delegate = context.delegate; if (delegate) { var delegateResult = maybeInvokeDelegate(delegate, context); if (delegateResult) { if (delegateResult === ContinueSentinel) continue; return delegateResult; } } if (\"next\" === context.method) context.sent = context._sent = context.arg;else if (\"throw\" === context.method) { if (\"suspendedStart\" === state) throw state = \"completed\", context.arg; context.dispatchException(context.arg); } else \"return\" === context.method && context.abrupt(\"return\", context.arg); state = \"executing\"; var record = tryCatch(innerFn, self, context); if (\"normal\" === record.type) { if (state = context.done ? \"completed\" : \"suspendedYield\", record.arg === ContinueSentinel) continue; return { value: record.arg, done: context.done }; } \"throw\" === record.type && (state = \"completed\", context.method = \"throw\", context.arg = record.arg); } }; } function maybeInvokeDelegate(delegate, context) { var methodName = context.method, method = delegate.iterator[methodName]; if (undefined === method) return context.delegate = null, \"throw\" === methodName && delegate.iterator[\"return\"] && (context.method = \"return\", context.arg = undefined, maybeInvokeDelegate(delegate, context), \"throw\" === context.method) || \"return\" !== methodName && (context.method = \"throw\", context.arg = new TypeError(\"The iterator does not provide a '\" + methodName + \"' method\")), ContinueSentinel; var record = tryCatch(method, delegate.iterator, context.arg); if (\"throw\" === record.type) return context.method = \"throw\", context.arg = record.arg, context.delegate = null, ContinueSentinel; var info = record.arg; return info ? info.done ? (context[delegate.resultName] = info.value, context.next = delegate.nextLoc, \"return\" !== context.method && (context.method = \"next\", context.arg = undefined), context.delegate = null, ContinueSentinel) : info : (context.method = \"throw\", context.arg = new TypeError(\"iterator result is not an object\"), context.delegate = null, ContinueSentinel); } function pushTryEntry(locs) { var entry = { tryLoc: locs[0] }; 1 in locs && (entry.catchLoc = locs[1]), 2 in locs && (entry.finallyLoc = locs[2], entry.afterLoc = locs[3]), this.tryEntries.push(entry); } function resetTryEntry(entry) { var record = entry.completion || {}; record.type = \"normal\", delete record.arg, entry.completion = record; } function Context(tryLocsList) { this.tryEntries = [{ tryLoc: \"root\" }], tryLocsList.forEach(pushTryEntry, this), this.reset(!0); } function values(iterable) { if (iterable || \"\" === iterable) { var iteratorMethod = iterable[iteratorSymbol]; if (iteratorMethod) return iteratorMethod.call(iterable); if (\"function\" == typeof iterable.next) return iterable; if (!isNaN(iterable.length)) { var i = -1, next = function next() { for (; ++i < iterable.length;) if (hasOwn.call(iterable, i)) return next.value = iterable[i], next.done = !1, next; return next.value = undefined, next.done = !0, next; }; return next.next = next; } } throw new TypeError(_typeof(iterable) + \" is not iterable\"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, defineProperty(Gp, \"constructor\", { value: GeneratorFunctionPrototype, configurable: !0 }), defineProperty(GeneratorFunctionPrototype, \"constructor\", { value: GeneratorFunction, configurable: !0 }), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, toStringTagSymbol, \"GeneratorFunction\"), exports.isGeneratorFunction = function (genFun) { var ctor = \"function\" == typeof genFun && genFun.constructor; return !!ctor && (ctor === GeneratorFunction || \"GeneratorFunction\" === (ctor.displayName || ctor.name)); }, exports.mark = function (genFun) { return Object.setPrototypeOf ? Object.setPrototypeOf(genFun, GeneratorFunctionPrototype) : (genFun.__proto__ = GeneratorFunctionPrototype, define(genFun, toStringTagSymbol, \"GeneratorFunction\")), genFun.prototype = Object.create(Gp), genFun; }, exports.awrap = function (arg) { return { __await: arg }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, asyncIteratorSymbol, function () { return this; }), exports.AsyncIterator = AsyncIterator, exports.async = function (innerFn, outerFn, self, tryLocsList, PromiseImpl) { void 0 === PromiseImpl && (PromiseImpl = Promise); var iter = new AsyncIterator(wrap(innerFn, outerFn, self, tryLocsList), PromiseImpl); return exports.isGeneratorFunction(outerFn) ? iter : iter.next().then(function (result) { return result.done ? result.value : iter.next(); }); }, defineIteratorMethods(Gp), define(Gp, toStringTagSymbol, \"Generator\"), define(Gp, iteratorSymbol, function () { return this; }), define(Gp, \"toString\", function () { return \"[object Generator]\"; }), exports.keys = function (val) { var object = Object(val), keys = []; for (var key in object) keys.push(key); return keys.reverse(), function next() { for (; keys.length;) { var key = keys.pop(); if (key in object) return next.value = key, next.done = !1, next; } return next.done = !0, next; }; }, exports.values = values, Context.prototype = { constructor: Context, reset: function reset(skipTempReset) { if (this.prev = 0, this.next = 0, this.sent = this._sent = undefined, this.done = !1, this.delegate = null, this.method = \"next\", this.arg = undefined, this.tryEntries.forEach(resetTryEntry), !skipTempReset) for (var name in this) \"t\" === name.charAt(0) && hasOwn.call(this, name) && !isNaN(+name.slice(1)) && (this[name] = undefined); }, stop: function stop() { this.done = !0; var rootRecord = this.tryEntries[0].completion; if (\"throw\" === rootRecord.type) throw rootRecord.arg; return this.rval; }, dispatchException: function dispatchException(exception) { if (this.done) throw exception; var context = this; function handle(loc, caught) { return record.type = \"throw\", record.arg = exception, context.next = loc, caught && (context.method = \"next\", context.arg = undefined), !!caught; } for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i], record = entry.completion; if (\"root\" === entry.tryLoc) return handle(\"end\"); if (entry.tryLoc <= this.prev) { var hasCatch = hasOwn.call(entry, \"catchLoc\"), hasFinally = hasOwn.call(entry, \"finallyLoc\"); if (hasCatch && hasFinally) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } else if (hasCatch) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); } else { if (!hasFinally) throw new Error(\"try statement without catch or finally\"); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } } } }, abrupt: function abrupt(type, arg) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc <= this.prev && hasOwn.call(entry, \"finallyLoc\") && this.prev < entry.finallyLoc) { var finallyEntry = entry; break; } } finallyEntry && (\"break\" === type || \"continue\" === type) && finallyEntry.tryLoc <= arg && arg <= finallyEntry.finallyLoc && (finallyEntry = null); var record = finallyEntry ? finallyEntry.completion : {}; return record.type = type, record.arg = arg, finallyEntry ? (this.method = \"next\", this.next = finallyEntry.finallyLoc, ContinueSentinel) : this.complete(record); }, complete: function complete(record, afterLoc) { if (\"throw\" === record.type) throw record.arg; return \"break\" === record.type || \"continue\" === record.type ? this.next = record.arg : \"return\" === record.type ? (this.rval = this.arg = record.arg, this.method = \"return\", this.next = \"end\") : \"normal\" === record.type && afterLoc && (this.next = afterLoc), ContinueSentinel; }, finish: function finish(finallyLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.finallyLoc === finallyLoc) return this.complete(entry.completion, entry.afterLoc), resetTryEntry(entry), ContinueSentinel; } }, \"catch\": function _catch(tryLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc === tryLoc) { var record = entry.completion; if (\"throw\" === record.type) { var thrown = record.arg; resetTryEntry(entry); } return thrown; } } throw new Error(\"illegal catch attempt\"); }, delegateYield: function delegateYield(iterable, resultName, nextLoc) { return this.delegate = { iterator: values(iterable), resultName: resultName, nextLoc: nextLoc }, \"next\" === this.method && (this.arg = undefined), ContinueSentinel; } }, exports; }\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); enumerableOnly && (symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; })), keys.push.apply(keys, symbols); } return keys; }\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = null != arguments[i] ? arguments[i] : {}; i % 2 ? ownKeys(Object(source), !0).forEach(function (key) { _defineProperty(target, key, source[key]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)) : ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } return target; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _toPropertyKey(arg) { var key = _toPrimitive(arg, \"string\"); return _typeof(key) === \"symbol\" ? key : String(key); }\nfunction _toPrimitive(input, hint) { if (_typeof(input) !== \"object\" || input === null) return input; var prim = input[Symbol.toPrimitive]; if (prim !== undefined) { var res = prim.call(input, hint || \"default\"); if (_typeof(res) !== \"object\") return res; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (hint === \"string\" ? String : Number)(input); }\nfunction asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { Promise.resolve(value).then(_next, _throw); } }\nfunction _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"next\", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"throw\", err); } _next(undefined); }); }; }\n\n\n\n\n\n\nvar EDIT_FORM_RECEIVER_PHONE_ID = 'bash--input-edit-adress-form-receiverPhone';\nvar Heading = function Heading() {\n  return (/* html */\"\\n    <div class=\\\"bash--heading\\\">\\n      <h3>Address Details</h3>\\n      <a href=\\\"#\\\" class=\\\"back-button--select\\\" data-view=\\\"select-address\\\">&lt; Back</a>\\n    </div>\\n\"\n  );\n};\nvar SaveButton = function SaveButton() {\n  return (/* html */\"\\n  <button\\n    class=\\\"submit btn-go-to-payment btn btn-large btn-success\\\"\\n    id=\\\"btn-save-address\\\"\\n    type=\\\"submit\\\">\\n    Save Address\\n  </button>\\n\"\n  );\n};\nvar EditAddressForm = function EditAddressForm(data) {\n  var _data$receiverName, _data$receiverPhone;\n  var fields = [{\n    name: 'addressId',\n    type: 'hidden',\n    value: data.id,\n    required: false\n  }, {\n    name: 'addressName',\n    type: 'hidden',\n    value: data.addressName,\n    required: false,\n    maxLength: 50\n  }, {\n    name: 'receiverName',\n    label: 'Recipient’s name',\n    required: true,\n    value: (_data$receiverName = data === null || data === void 0 ? void 0 : data.receiverName) !== null && _data$receiverName !== void 0 ? _data$receiverName : ''\n  }, {\n    name: 'receiverPhone',\n    label: 'Recipient’s mobile number',\n    required: true,\n    type: 'tel',\n    value: (_data$receiverPhone = data === null || data === void 0 ? void 0 : data.receiverPhone) !== null && _data$receiverPhone !== void 0 ? _data$receiverPhone : '',\n    helperText: 'We send shipping updates to this number.',\n    minlength: 9,\n    error: 'Please enter a valid phone number',\n    containerClasses: 'custom-field-complement',\n    // for sa flag\n    idOverride: EDIT_FORM_RECEIVER_PHONE_ID\n  }];\n  return (/* html */\"\\n    \".concat(Heading(), \"\\n    \").concat((0,_ContactCard__WEBPACK_IMPORTED_MODULE_3__[\"default\"])(data), \"\\n    <form id=\\\"bash--edit-address-form\\\" method=\\\"post\\\">\\n      \").concat(fields.map(function (field) {\n      return (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_4__[\"default\"])(field);\n    }).join(''), \"\\n      \").concat(SaveButton(), \"\\n    </form>\\n  \")\n  );\n};\nvar submitEditAddressForm = /*#__PURE__*/function () {\n  var _ref = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee2(event) {\n    var form, formData, addressId, addressName, receiverName, receiverPhone, invalidFields;\n    return _regeneratorRuntime().wrap(function _callee2$(_context2) {\n      while (1) switch (_context2.prev = _context2.next) {\n        case 0:\n          event.preventDefault();\n\n          // PULL ALL FORM FIELDS\n          form = document.getElementById('bash--edit-address-form');\n          formData = new FormData(form);\n          addressId = formData.get('addressId');\n          addressName = formData.get('addressName');\n          receiverName = formData.get('receiverName');\n          receiverPhone = formData.get('receiverPhone');\n          receiverPhone = (0,_utils_phoneFields__WEBPACK_IMPORTED_MODULE_0__.formatPhoneNumber)(receiverPhone, 'ZA').trim();\n\n          // VALIDATE FIELDS\n          invalidFields = [];\n          if (!receiverName) invalidFields.push('receiverName');\n          if (!receiverPhone || !(0,libphonenumber_js__WEBPACK_IMPORTED_MODULE_5__.isValidNumber)(receiverPhone, 'ZA')) {\n            invalidFields.push('receiverPhone');\n            $(\"#\".concat(EDIT_FORM_RECEIVER_PHONE_ID)).addClass('invalid');\n          }\n\n          // APPLY VALIDATION UI\n          if (!(invalidFields.length > 0)) {\n            _context2.next = 17;\n            break;\n          }\n          console.error({\n            invalidFields: invalidFields\n          });\n          $('#bash--edit-address-form').addClass('show-form-errors');\n          $(\"#bash--input-\".concat(invalidFields[0])).focus();\n          window.postMessage({\n            type: 'ADDRESS_VALIDATION_ERROR',\n            message: 'Address validation error. See invalidFields.',\n            invalidFields: invalidFields\n          }, '*');\n          return _context2.abrupt(\"return\");\n        case 17:\n          // POST ADDRESS UPDATE AND CHANGE VIEW\n          (0,_utils_services__WEBPACK_IMPORTED_MODULE_1__.getAddressByName)(addressName).then( /*#__PURE__*/function () {\n            var _ref2 = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee(address) {\n              var payload, setAddressResponse, success;\n              return _regeneratorRuntime().wrap(function _callee$(_context) {\n                while (1) switch (_context.prev = _context.next) {\n                  case 0:\n                    payload = _objectSpread(_objectSpread({}, address), {}, {\n                      addressId: addressId,\n                      addressName: addressName,\n                      receiverName: receiverName,\n                      receiverPhone: receiverPhone,\n                      geoCoordinates: (address === null || address === void 0 ? void 0 : address.geoCoordinate) || [] // for shippingData\n                    }); // Apply the selected address to customers orderForm.\n                    _context.next = 3;\n                    return (0,_utils_setAddress__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(payload, {\n                      validateExtraFields: false\n                    });\n                  case 3:\n                    setAddressResponse = _context.sent;\n                    success = setAddressResponse.success;\n                    if (success) {\n                      _context.next = 8;\n                      break;\n                    }\n                    console.error('Set address error', {\n                      setAddressResponse: setAddressResponse\n                    });\n                    return _context.abrupt(\"return\");\n                  case 8:\n                    (0,_utils_services__WEBPACK_IMPORTED_MODULE_1__.addOrUpdateAddress)(payload);\n                    window.postMessage({\n                      action: 'setDeliveryView',\n                      view: 'select-address'\n                    });\n                  case 10:\n                  case \"end\":\n                    return _context.stop();\n                }\n              }, _callee);\n            }));\n            return function (_x2) {\n              return _ref2.apply(this, arguments);\n            };\n          }());\n        case 18:\n        case \"end\":\n          return _context2.stop();\n      }\n    }, _callee2);\n  }));\n  return function submitEditAddressForm(_x) {\n    return _ref.apply(this, arguments);\n  };\n}();\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (EditAddressForm);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/EditAddressForm.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"EDIT_FORM_RECEIVER_PHONE_ID\": () => (/* binding */ EDIT_FORM_RECEIVER_PHONE_ID),\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__),\n/* harmony export */   \"submitEditAddressForm\": () => (/* binding */ submitEditAddressForm)\n/* harmony export */ });\n/* harmony import */ var libphonenumber_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! libphonenumber-js */ \"./node_modules/libphonenumber-js/index.es6.exports/isValidNumber.js\");\n/* harmony import */ var _utils_phoneFields__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/phoneFields */ \"./src/utils/phoneFields.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _utils_setAddress__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/setAddress */ \"./src/utils/setAddress.js\");\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\n/* harmony import */ var _FormComponents__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./FormComponents */ \"./src/partials/Deliver/FormComponents.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\nfunction _regeneratorRuntime() { \"use strict\"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return exports; }; var exports = {}, Op = Object.prototype, hasOwn = Op.hasOwnProperty, defineProperty = Object.defineProperty || function (obj, key, desc) { obj[key] = desc.value; }, $Symbol = \"function\" == typeof Symbol ? Symbol : {}, iteratorSymbol = $Symbol.iterator || \"@@iterator\", asyncIteratorSymbol = $Symbol.asyncIterator || \"@@asyncIterator\", toStringTagSymbol = $Symbol.toStringTag || \"@@toStringTag\"; function define(obj, key, value) { return Object.defineProperty(obj, key, { value: value, enumerable: !0, configurable: !0, writable: !0 }), obj[key]; } try { define({}, \"\"); } catch (err) { define = function define(obj, key, value) { return obj[key] = value; }; } function wrap(innerFn, outerFn, self, tryLocsList) { var protoGenerator = outerFn && outerFn.prototype instanceof Generator ? outerFn : Generator, generator = Object.create(protoGenerator.prototype), context = new Context(tryLocsList || []); return defineProperty(generator, \"_invoke\", { value: makeInvokeMethod(innerFn, self, context) }), generator; } function tryCatch(fn, obj, arg) { try { return { type: \"normal\", arg: fn.call(obj, arg) }; } catch (err) { return { type: \"throw\", arg: err }; } } exports.wrap = wrap; var ContinueSentinel = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var IteratorPrototype = {}; define(IteratorPrototype, iteratorSymbol, function () { return this; }); var getProto = Object.getPrototypeOf, NativeIteratorPrototype = getProto && getProto(getProto(values([]))); NativeIteratorPrototype && NativeIteratorPrototype !== Op && hasOwn.call(NativeIteratorPrototype, iteratorSymbol) && (IteratorPrototype = NativeIteratorPrototype); var Gp = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(IteratorPrototype); function defineIteratorMethods(prototype) { [\"next\", \"throw\", \"return\"].forEach(function (method) { define(prototype, method, function (arg) { return this._invoke(method, arg); }); }); } function AsyncIterator(generator, PromiseImpl) { function invoke(method, arg, resolve, reject) { var record = tryCatch(generator[method], generator, arg); if (\"throw\" !== record.type) { var result = record.arg, value = result.value; return value && \"object\" == _typeof(value) && hasOwn.call(value, \"__await\") ? PromiseImpl.resolve(value.__await).then(function (value) { invoke(\"next\", value, resolve, reject); }, function (err) { invoke(\"throw\", err, resolve, reject); }) : PromiseImpl.resolve(value).then(function (unwrapped) { result.value = unwrapped, resolve(result); }, function (error) { return invoke(\"throw\", error, resolve, reject); }); } reject(record.arg); } var previousPromise; defineProperty(this, \"_invoke\", { value: function value(method, arg) { function callInvokeWithMethodAndArg() { return new PromiseImpl(function (resolve, reject) { invoke(method, arg, resolve, reject); }); } return previousPromise = previousPromise ? previousPromise.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); } }); } function makeInvokeMethod(innerFn, self, context) { var state = \"suspendedStart\"; return function (method, arg) { if (\"executing\" === state) throw new Error(\"Generator is already running\"); if (\"completed\" === state) { if (\"throw\" === method) throw arg; return { value: void 0, done: !0 }; } for (context.method = method, context.arg = arg;;) { var delegate = context.delegate; if (delegate) { var delegateResult = maybeInvokeDelegate(delegate, context); if (delegateResult) { if (delegateResult === ContinueSentinel) continue; return delegateResult; } } if (\"next\" === context.method) context.sent = context._sent = context.arg;else if (\"throw\" === context.method) { if (\"suspendedStart\" === state) throw state = \"completed\", context.arg; context.dispatchException(context.arg); } else \"return\" === context.method && context.abrupt(\"return\", context.arg); state = \"executing\"; var record = tryCatch(innerFn, self, context); if (\"normal\" === record.type) { if (state = context.done ? \"completed\" : \"suspendedYield\", record.arg === ContinueSentinel) continue; return { value: record.arg, done: context.done }; } \"throw\" === record.type && (state = \"completed\", context.method = \"throw\", context.arg = record.arg); } }; } function maybeInvokeDelegate(delegate, context) { var methodName = context.method, method = delegate.iterator[methodName]; if (undefined === method) return context.delegate = null, \"throw\" === methodName && delegate.iterator[\"return\"] && (context.method = \"return\", context.arg = undefined, maybeInvokeDelegate(delegate, context), \"throw\" === context.method) || \"return\" !== methodName && (context.method = \"throw\", context.arg = new TypeError(\"The iterator does not provide a '\" + methodName + \"' method\")), ContinueSentinel; var record = tryCatch(method, delegate.iterator, context.arg); if (\"throw\" === record.type) return context.method = \"throw\", context.arg = record.arg, context.delegate = null, ContinueSentinel; var info = record.arg; return info ? info.done ? (context[delegate.resultName] = info.value, context.next = delegate.nextLoc, \"return\" !== context.method && (context.method = \"next\", context.arg = undefined), context.delegate = null, ContinueSentinel) : info : (context.method = \"throw\", context.arg = new TypeError(\"iterator result is not an object\"), context.delegate = null, ContinueSentinel); } function pushTryEntry(locs) { var entry = { tryLoc: locs[0] }; 1 in locs && (entry.catchLoc = locs[1]), 2 in locs && (entry.finallyLoc = locs[2], entry.afterLoc = locs[3]), this.tryEntries.push(entry); } function resetTryEntry(entry) { var record = entry.completion || {}; record.type = \"normal\", delete record.arg, entry.completion = record; } function Context(tryLocsList) { this.tryEntries = [{ tryLoc: \"root\" }], tryLocsList.forEach(pushTryEntry, this), this.reset(!0); } function values(iterable) { if (iterable || \"\" === iterable) { var iteratorMethod = iterable[iteratorSymbol]; if (iteratorMethod) return iteratorMethod.call(iterable); if (\"function\" == typeof iterable.next) return iterable; if (!isNaN(iterable.length)) { var i = -1, next = function next() { for (; ++i < iterable.length;) if (hasOwn.call(iterable, i)) return next.value = iterable[i], next.done = !1, next; return next.value = undefined, next.done = !0, next; }; return next.next = next; } } throw new TypeError(_typeof(iterable) + \" is not iterable\"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, defineProperty(Gp, \"constructor\", { value: GeneratorFunctionPrototype, configurable: !0 }), defineProperty(GeneratorFunctionPrototype, \"constructor\", { value: GeneratorFunction, configurable: !0 }), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, toStringTagSymbol, \"GeneratorFunction\"), exports.isGeneratorFunction = function (genFun) { var ctor = \"function\" == typeof genFun && genFun.constructor; return !!ctor && (ctor === GeneratorFunction || \"GeneratorFunction\" === (ctor.displayName || ctor.name)); }, exports.mark = function (genFun) { return Object.setPrototypeOf ? Object.setPrototypeOf(genFun, GeneratorFunctionPrototype) : (genFun.__proto__ = GeneratorFunctionPrototype, define(genFun, toStringTagSymbol, \"GeneratorFunction\")), genFun.prototype = Object.create(Gp), genFun; }, exports.awrap = function (arg) { return { __await: arg }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, asyncIteratorSymbol, function () { return this; }), exports.AsyncIterator = AsyncIterator, exports.async = function (innerFn, outerFn, self, tryLocsList, PromiseImpl) { void 0 === PromiseImpl && (PromiseImpl = Promise); var iter = new AsyncIterator(wrap(innerFn, outerFn, self, tryLocsList), PromiseImpl); return exports.isGeneratorFunction(outerFn) ? iter : iter.next().then(function (result) { return result.done ? result.value : iter.next(); }); }, defineIteratorMethods(Gp), define(Gp, toStringTagSymbol, \"Generator\"), define(Gp, iteratorSymbol, function () { return this; }), define(Gp, \"toString\", function () { return \"[object Generator]\"; }), exports.keys = function (val) { var object = Object(val), keys = []; for (var key in object) keys.push(key); return keys.reverse(), function next() { for (; keys.length;) { var key = keys.pop(); if (key in object) return next.value = key, next.done = !1, next; } return next.done = !0, next; }; }, exports.values = values, Context.prototype = { constructor: Context, reset: function reset(skipTempReset) { if (this.prev = 0, this.next = 0, this.sent = this._sent = undefined, this.done = !1, this.delegate = null, this.method = \"next\", this.arg = undefined, this.tryEntries.forEach(resetTryEntry), !skipTempReset) for (var name in this) \"t\" === name.charAt(0) && hasOwn.call(this, name) && !isNaN(+name.slice(1)) && (this[name] = undefined); }, stop: function stop() { this.done = !0; var rootRecord = this.tryEntries[0].completion; if (\"throw\" === rootRecord.type) throw rootRecord.arg; return this.rval; }, dispatchException: function dispatchException(exception) { if (this.done) throw exception; var context = this; function handle(loc, caught) { return record.type = \"throw\", record.arg = exception, context.next = loc, caught && (context.method = \"next\", context.arg = undefined), !!caught; } for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i], record = entry.completion; if (\"root\" === entry.tryLoc) return handle(\"end\"); if (entry.tryLoc <= this.prev) { var hasCatch = hasOwn.call(entry, \"catchLoc\"), hasFinally = hasOwn.call(entry, \"finallyLoc\"); if (hasCatch && hasFinally) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } else if (hasCatch) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); } else { if (!hasFinally) throw new Error(\"try statement without catch or finally\"); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } } } }, abrupt: function abrupt(type, arg) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc <= this.prev && hasOwn.call(entry, \"finallyLoc\") && this.prev < entry.finallyLoc) { var finallyEntry = entry; break; } } finallyEntry && (\"break\" === type || \"continue\" === type) && finallyEntry.tryLoc <= arg && arg <= finallyEntry.finallyLoc && (finallyEntry = null); var record = finallyEntry ? finallyEntry.completion : {}; return record.type = type, record.arg = arg, finallyEntry ? (this.method = \"next\", this.next = finallyEntry.finallyLoc, ContinueSentinel) : this.complete(record); }, complete: function complete(record, afterLoc) { if (\"throw\" === record.type) throw record.arg; return \"break\" === record.type || \"continue\" === record.type ? this.next = record.arg : \"return\" === record.type ? (this.rval = this.arg = record.arg, this.method = \"return\", this.next = \"end\") : \"normal\" === record.type && afterLoc && (this.next = afterLoc), ContinueSentinel; }, finish: function finish(finallyLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.finallyLoc === finallyLoc) return this.complete(entry.completion, entry.afterLoc), resetTryEntry(entry), ContinueSentinel; } }, \"catch\": function _catch(tryLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc === tryLoc) { var record = entry.completion; if (\"throw\" === record.type) { var thrown = record.arg; resetTryEntry(entry); } return thrown; } } throw new Error(\"illegal catch attempt\"); }, delegateYield: function delegateYield(iterable, resultName, nextLoc) { return this.delegate = { iterator: values(iterable), resultName: resultName, nextLoc: nextLoc }, \"next\" === this.method && (this.arg = undefined), ContinueSentinel; } }, exports; }\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); enumerableOnly && (symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; })), keys.push.apply(keys, symbols); } return keys; }\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = null != arguments[i] ? arguments[i] : {}; i % 2 ? ownKeys(Object(source), !0).forEach(function (key) { _defineProperty(target, key, source[key]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)) : ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } return target; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _toPropertyKey(arg) { var key = _toPrimitive(arg, \"string\"); return _typeof(key) === \"symbol\" ? key : String(key); }\nfunction _toPrimitive(input, hint) { if (_typeof(input) !== \"object\" || input === null) return input; var prim = input[Symbol.toPrimitive]; if (prim !== undefined) { var res = prim.call(input, hint || \"default\"); if (_typeof(res) !== \"object\") return res; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (hint === \"string\" ? String : Number)(input); }\nfunction asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { Promise.resolve(value).then(_next, _throw); } }\nfunction _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"next\", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"throw\", err); } _next(undefined); }); }; }\n\n\n\n\n\n\nvar EDIT_FORM_RECEIVER_PHONE_ID = 'bash--input-edit-adress-form-receiverPhone';\nvar EditAddressForm = function EditAddressForm(data) {\n  var _data$receiverName, _data$receiverPhone;\n  var fields = [{\n    name: 'addressId',\n    type: 'hidden',\n    value: data.id,\n    required: false\n  }, {\n    name: 'addressName',\n    type: 'hidden',\n    value: data.addressName,\n    required: false,\n    maxLength: 50\n  }, {\n    name: 'receiverName',\n    label: 'Recipient’s name',\n    required: true,\n    value: (_data$receiverName = data === null || data === void 0 ? void 0 : data.receiverName) !== null && _data$receiverName !== void 0 ? _data$receiverName : ''\n  }, {\n    name: 'receiverPhone',\n    label: 'Recipient’s mobile number',\n    required: true,\n    type: 'tel',\n    value: (_data$receiverPhone = data === null || data === void 0 ? void 0 : data.receiverPhone) !== null && _data$receiverPhone !== void 0 ? _data$receiverPhone : '',\n    helperText: 'We send shipping updates to this number.',\n    minlength: 9,\n    error: 'Please enter a valid phone number',\n    containerClasses: 'custom-field-complement',\n    // for sa flag\n    idOverride: EDIT_FORM_RECEIVER_PHONE_ID\n  }];\n  return (/* html */\"\\n    \".concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_4__.AddressSectionHeading)('Address Details'), \"\\n    \").concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_4__.ContactCard)(data), \"\\n    <form id=\\\"bash--edit-address-form\\\" method=\\\"post\\\">\\n      \").concat(fields.map(function (field) {\n      return (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_3__[\"default\"])(field);\n    }).join(''), \"\\n      \").concat((0,_FormComponents__WEBPACK_IMPORTED_MODULE_4__.SubmitButton)(), \"\\n    </form>\\n  \")\n  );\n};\nvar submitEditAddressForm = /*#__PURE__*/function () {\n  var _ref = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee2(event) {\n    var form, formData, addressId, addressName, receiverName, receiverPhone, invalidFields;\n    return _regeneratorRuntime().wrap(function _callee2$(_context2) {\n      while (1) switch (_context2.prev = _context2.next) {\n        case 0:\n          event.preventDefault();\n\n          // PULL ALL FORM FIELDS\n          form = document.getElementById('bash--edit-address-form');\n          formData = new FormData(form);\n          addressId = formData.get('addressId');\n          addressName = formData.get('addressName');\n          receiverName = formData.get('receiverName');\n          receiverPhone = formData.get('receiverPhone');\n          receiverPhone = (0,_utils_phoneFields__WEBPACK_IMPORTED_MODULE_0__.formatPhoneNumber)(receiverPhone, 'ZA').trim();\n\n          // VALIDATE FIELDS\n          invalidFields = [];\n          if (!receiverName) invalidFields.push('receiverName');\n          if (!receiverPhone || !(0,libphonenumber_js__WEBPACK_IMPORTED_MODULE_5__.isValidNumber)(receiverPhone, 'ZA')) {\n            invalidFields.push('receiverPhone');\n            $(\"#\".concat(EDIT_FORM_RECEIVER_PHONE_ID)).addClass('invalid');\n          }\n\n          // APPLY VALIDATION UI\n          if (!(invalidFields.length > 0)) {\n            _context2.next = 17;\n            break;\n          }\n          console.error({\n            invalidFields: invalidFields\n          });\n          $('#bash--edit-address-form').addClass('show-form-errors');\n          $(\"#bash--input-\".concat(invalidFields[0])).focus();\n          window.postMessage({\n            type: 'ADDRESS_VALIDATION_ERROR',\n            message: 'Address validation error. See invalidFields.',\n            invalidFields: invalidFields\n          }, '*');\n          return _context2.abrupt(\"return\");\n        case 17:\n          // POST ADDRESS UPDATE AND CHANGE VIEW\n          (0,_utils_services__WEBPACK_IMPORTED_MODULE_1__.getAddressByName)(addressName).then( /*#__PURE__*/function () {\n            var _ref2 = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee(address) {\n              var payload, setAddressResponse, success;\n              return _regeneratorRuntime().wrap(function _callee$(_context) {\n                while (1) switch (_context.prev = _context.next) {\n                  case 0:\n                    payload = _objectSpread(_objectSpread({}, address), {}, {\n                      addressId: addressId,\n                      addressName: addressName,\n                      receiverName: receiverName,\n                      receiverPhone: receiverPhone,\n                      geoCoordinates: (address === null || address === void 0 ? void 0 : address.geoCoordinate) || [] // for shippingData\n                    }); // Apply the selected address to customers orderForm.\n                    _context.next = 3;\n                    return (0,_utils_setAddress__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(payload, {\n                      validateExtraFields: false\n                    });\n                  case 3:\n                    setAddressResponse = _context.sent;\n                    success = setAddressResponse.success;\n                    if (success) {\n                      _context.next = 8;\n                      break;\n                    }\n                    console.error('Set address error', {\n                      setAddressResponse: setAddressResponse\n                    });\n                    return _context.abrupt(\"return\");\n                  case 8:\n                    (0,_utils_services__WEBPACK_IMPORTED_MODULE_1__.addOrUpdateAddress)(payload);\n                    window.postMessage({\n                      action: 'setDeliveryView',\n                      view: 'select-address'\n                    });\n                  case 10:\n                  case \"end\":\n                    return _context.stop();\n                }\n              }, _callee);\n            }));\n            return function (_x2) {\n              return _ref2.apply(this, arguments);\n            };\n          }());\n        case 18:\n        case \"end\":\n          return _context2.stop();\n      }\n    }, _callee2);\n  }));\n  return function submitEditAddressForm(_x) {\n    return _ref.apply(this, arguments);\n  };\n}();\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (EditAddressForm);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/EditAddressForm.js?");
 
 /***/ }),
 
@@ -427,6 +417,16 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _RICAForm__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./RICAForm */ \"./src/partials/Deliver/RICAForm.js\");\n/* harmony import */ var _TVLicenseForm__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./TVLicenseForm */ \"./src/partials/Deliver/TVLicenseForm.js\");\n\n\nvar ExtraFieldsContainer = function ExtraFieldsContainer(_ref) {\n  var hasTV = _ref.hasTV,\n    hasSim = _ref.hasSim;\n  var showTVLicenseForm = \"\\n    <div id=\\\"tv-license-form\\\">\\n      <hr>\\n      <div class=\\\"bash--heading sub-heading heading-with-description\\\">\\n        <h3>TV license information needed</h3>\\n        <p class=\\\"tfg-custom-subtitle\\\">Please provide your ID number to validate your TV Licence.</p>\\n      </div>\\n      \".concat((0,_TVLicenseForm__WEBPACK_IMPORTED_MODULE_1__[\"default\"])(), \"\\n    </div>\\n  \");\n  var showRICAForm = \"\\n    <div id=\\\"rica-form\\\">\\n      <hr>\\n      <div class=\\\"bash--heading sub-heading heading-with-description\\\">\\n        <h3>Rica information required</h3>\\n        <p class=\\\"tfg-custom-subtitle\\\">\\n          To RICA your SIM card, provide your SA ID (or foreign passport) number and your address as\\n          it appears on a valid proof of residence.\\n        </p> \\n      </div>\\n        \".concat((0,_RICAForm__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(), \"\\n    </div>\\n    \");\n  return \"\\n  <section class=\\\"bash--extra-fields bash--delivery-view\\\" data-section=\\\"select-address\\\">\\n    \".concat(hasTV ? showTVLicenseForm : '', \"\\n    \").concat(hasSim ? showRICAForm : '', \"\\n  </section>\");\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (ExtraFieldsContainer);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/ExtraFieldsContainer.js?");
+
+/***/ }),
+
+/***/ "./src/partials/Deliver/FormComponents.js":
+/*!************************************************!*\
+  !*** ./src/partials/Deliver/FormComponents.js ***!
+  \************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"AddressSectionHeading\": () => (/* binding */ AddressSectionHeading),\n/* harmony export */   \"ContactCard\": () => (/* binding */ ContactCard),\n/* harmony export */   \"SubmitButton\": () => (/* binding */ SubmitButton)\n/* harmony export */ });\n/**\n * ContactCard - View Component of a Address Summary\n * @param {Object} data - Address data\n */\nvar ContactCard = function ContactCard(data) {\n  var number = data.number,\n    businessName = data.businessName,\n    street = data.street,\n    neighborhood = data.neighborhood,\n    city = data.city,\n    postalCode = data.postalCode;\n  var addressLine = [businessName, \"\".concat(number ? \"\".concat(number, \" \") : '').concat(street), neighborhood !== null && neighborhood !== void 0 ? neighborhood : city, postalCode].filter(function (item) {\n    return item !== undefined && item !== null && item !== '';\n  }).join(', ').trim();\n  return (/* html */\"\\n    <div style=\\\"border: 0.5px solid #0404041A; border-radius: 8px; padding: 16px; display: flex; align-items: center;\\\">\\n      <svg width=\\\"24\\\" height=\\\"19\\\" viewBox=\\\"0 0 24 19\\\" fill=\\\"none\\\" xmlns=\\\"http://www.w3.org/2000/svg\\\">\\n        <path d=\\\"M1.5 18.2496H22.5M14.25 18.2496V12.2496H9.75V18.2496M3.75 8.9393V18.2496M20.25 18.2496V8.9393M2.25 10.4393L11.4694 1.21899C11.539 1.14926 11.6217 1.09394 11.7128 1.05619C11.8038 1.01845 11.9014 0.999023 12 0.999023C12.0986 0.999023 12.1962 1.01845 12.2872 1.05619C12.3783 1.09394 12.461 1.14926 12.5306 1.21899L21.75 10.4393\\\" stroke=\\\"black\\\" stroke-width=\\\"1.5\\\" stroke-linecap=\\\"round\\\" stroke-linejoin=\\\"round\\\"/>\\n      </svg>\\n      <p style=\\\"margin-left: 16px; line-height: 1; font-size: 14px; margin-bottom:0;\\\">\".concat(addressLine, \"</p>\\n  </div>\\n  \")\n  );\n};\n\n/**\n * SubmitButton - View Component of a Submit Button\n * @returns {string} - HTML for a Submit Button\n\n */\nvar SubmitButton = function SubmitButton() {\n  return (/* html */\"\\n  <button\\n    class=\\\"submit btn-go-to-payment btn btn-large btn-success\\\"\\n    id=\\\"btn-save-address\\\"\\n    type=\\\"submit\\\">\\n      Save address\\n  </button>\\n\"\n  );\n};\n\n/**\n * AddressSectionHeading - View Component of a Heading\n * @param {string} title - Title of the Heading\n * @returns {string} - HTML for a Heading\n */\nvar AddressSectionHeading = function AddressSectionHeading(title) {\n  return (/* html */\"\\n  <div class=\\\"bash--heading\\\">\\n    <h3>\".concat(title, \"</h3>\\n    <a href=\\\"#\\\" class=\\\"back-button--select\\\" data-view=\\\"select-address\\\">&lt; Back</a>\\n  </div>\\n\")\n  );\n};\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/FormComponents.js?");
 
 /***/ }),
